+++ chessdotcom/client.py
@@ -181,3 +181,27 @@ class ChessDotComClient(Client):
             request_config["headers"]["User-Agent"] = user_agent
 
         self.request_config = request_config
+
+
+class Resource(object):
+    _base_url = "https://api.chess.com/pub"
+
+    def __init__(
+        self,
+        uri="",
+        top_level_attr=None,
+        no_json=False,
+        tts=0,
+        request_config=None,
+        times_requested=0,
+    ):
+        self.uri = uri
+        self.top_level_attr = top_level_attr
+        self.no_json = no_json
+        self._request_config = request_config or dict()
+        self.tts = tts
+        self.times_requested = times_requested
+
+    @property
+    def request_config(self):
+        return dict(url=self._base_url + self.uri, **self._request_config)
